[FEAT] data class 생성 및 errorResponse 분리 - base - baseResponse - authResponse
Fixes: TICO-133
