--- Mobile/PomoroDo/app/src/main/java/com/tico/pomorodo/data/model/Auth.kt
@@ -1,4 +1,0 @@
-package com.tico.pomorodo.data.model
-
-data class Auth(val status: Int, val message: String, val data: Token)
-data class Token(val accessToken: String)
