+++ Mobile/PomoroDo/app/src/main/java/com/tico/pomorodo/data/model/Base.kt
@@ -1,0 +1,3 @@
+package com.tico.pomorodo.data.model
+
+data class Base<T>(val status: Int, val message: String, val data: T)
