--- Mobile/PomoroDo/app/src/main/java/com/tico/pomorodo/data/remote/models/request/AuthRequestBody.kt
@@ -1,8 +1,0 @@
-package com.tico.pomorodo.data.remote.models.request
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class UserInfoRequestBody(
-    val nickname: String,
-)
