+++ Mobile/PomoroDo/app/src/main/java/com/tico/pomorodo/data/remote/models/response/ErrorResponse.kt
@@ -1,0 +1,10 @@
+package com.tico.pomorodo.data.remote.models.response
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ErrorResponse(
+    val status: Int,
+    val code: String,
+    val message: String
+)
